requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
groq==0.4.2
torch==2.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
playwright==1.41.0
pillow==10.2.0
pytesseract==0.3.10
